Rename ContactSheetBuilder to ContactSheetJob, slim down
