+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - **Footage analysis is dramatically faster.** Visual transcripts (LLM-driven, frame-by-frame) are gone. Each clip now gets a contact sheet (a single image grid covering the whole clip with timestamps burned in), a clean dialogue script (no timing weight), and the existing markdown summary. Contact sheet generation is pure ffmpeg — no LLM in that step — so analysis runtime drops from roughly 1× footage length to a small fraction of it. The roughcut agent reads the contact sheet to "see" a clip at a glance, the script for cheap dialogue context, and the audio transcript only when it needs word-level cut timing.
 - **`summarize-video` skill folded into `analyze-video`.** One skill, one parent dispatch, one sub-agent per clip.
-- **New `build_contact_sheets.rb` orchestrator.** Takes a library name plus an explicit list of clip filenames; generates `_full.jpg` for each (plus 10-minute chunk sheets for clips longer than 10 minutes) and updates library.yaml. Skips clips that already have a sheet. Single-threaded by design — the parent agent decides how many invocations to run in parallel based on machine headroom.
+- **New `contact_sheet_job.rb` orchestrator.** Takes a library name plus an explicit list of clip filenames; generates `_full.jpg` for each (plus 10-minute chunk sheets for clips longer than 10 minutes) and updates library.yaml. Always rebuilds for the clips it's given. Single-threaded by design — the parent agent decides how many invocations to run in parallel based on machine headroom.
 - **New `build_scripts.rb` orchestrator.** Takes a library name plus an explicit list of clip filenames; writes a clean dialogue script per clip from its audio transcript and updates library.yaml. Pure JSON parsing — no LLM. Skips clips whose script already exists. Single-threaded by design — parent agent parallelizes by launching multiple invocations.
 - **Analyze-video sub-agents now write each summary in one shot.** The four-placeholder skeleton + four-Edit dance is gone; the sub-agent reads the contact sheet and script, then issues a single `Write` with the full markdown. Drops per-clip from 7 tool calls to 3 — roughly halves the slowest-batch wall.
 - **Library schema:** new `script:` and `contact_sheet:` fields per video; `visual_transcript:` is deprecated. Old libraries keep their `visual_*.json` files on disk — no migration runs. Re-run analysis on an old library to populate the new fields.
